+++ aom_dsp/inv_txfm.h
@@ -55,13 +55,13 @@ static INLINE tran_high_t check_range(tran_high_t input, int bd) {
 #if CONFIG_MRC_TX
 // These each perform dct but add coefficients based on a mask
 void aom_imrc32x32_1024_add_c(const tran_low_t *input, uint8_t *dest,
-                              int stride, int *mask);
+                              int stride, uint8_t *mask);
 
 void aom_imrc32x32_135_add_c(const tran_low_t *input, uint8_t *dest, int stride,
-                             int *mask);
+                             uint8_t *mask);
 
 void aom_imrc32x32_34_add_c(const tran_low_t *input, uint8_t *dest, int stride,
-                            int *mask);
+                            uint8_t *mask);
 #endif  // CONFIG_MRC_TX
 
 void aom_idct4_c(const tran_low_t *input, tran_low_t *output);
